Move `conj-unique` to unmo.dictionary renamed with `add-sentence`
The conj-unique function has been used only in unmo.dictionary and
there's no reason to be defined in unmo.util.

In addition, conj-unique has returned a lazy-seq if the argument
coll was nil or another lazy-seq, however, it's unnecessary to be
a lazy for the data structure of the dictionary, so I defined a
new function named `add-sentence` similar to the `conj-unique` but
always returns a vector, converting lazy-seq if necessary.

Plus, `conj-unique` was renamed to `conj-with-distinct` because of
an explicity of what the function does... a bit maybe.